requirements.txt
# DesertNet-X Requirements
# Python 3.9+ / macOS M1 CPU
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
albumentations>=1.3.0
matplotlib>=3.7.0
opencv-python-headless>=4.7.0
